pyproject.toml
[project]
name = "sanspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "psycopg>=3.2.13",
    "pyqt6>=6.10.0",
    "python-dotenv>=1.2.1",
]
